requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
openai
langchain
langchain-community
langchain-openai
faiss-cpu
requests
langchain-core
langchain-experimental